feat: price 데이터 타입 BigDecimal -> int로 수정 / 잔액, 주문/결제 비즈니스 로직 생성 / 주문서 엔티티 추가

+++ src/main/java/com/hhplus/hhplus_week3_4_5/ecommerce/controller/product/dto/AddProductApiReqDto.java
@@ -4,7 +4,6 @@ import com.hhplus.hhplus_week3_4_5.ecommerce.domain.product.ProductEnums;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.List;
 
 public record AddProductApiReqDto(
@@ -13,7 +12,7 @@ public record AddProductApiReqDto(
         @Schema(description = "상품 타입 Enum")
         ProductEnums.Type type,
         @Schema(description = "상품 가격")
-        BigDecimal price,
+        int price,
         @Schema(description = "사용 여부")
         boolean useYn,
         @Schema(description = "상품 옵션 리스트")
@@ -34,7 +33,7 @@ public record AddProductApiReqDto(
                 @Schema(description = "상품 옵션 값")
                 String value,
                 @Schema(description = "상품 옵션 가격")
-                BigDecimal price,
+                int price,
                 @Schema(description = "사용 여부")
                 boolean useYn
         ) implements Serializable {
